Changed variables associated with coin to Decimal to allow for higher values, powerup costs rise exponentially now with a cap, added new powerups, updates settings page with new difficulty option, and various design changes

+++ Crypto Clicker/Views/CoinView.swift
@@ -10,7 +10,7 @@ import AudioToolbox
 
 struct CoinView: View {
     
-    @Binding var coinValue: Int
+    @Binding var coinValue: Decimal
     @ObservedObject var settings: SettingsModel // Observe changes in settings
     let incrementAction: () -> Void
     @State private var isCoinPressed = false // State to control animation
